Bootstrap release workflow
This commit create a new `dist` make target that packages a `tar.gz`
with a compatible layout with the current OSIE.

In this way it can work as quick replacement in tinkerbell/sandbox.

It is a manual process at this stage but the tar.gz gets pushed to s3
and it can be downloaded via sandbox. Right now take this URL as "good
enough":

http://s.gianarb.it/noname/noname-master.tar.gz

When you run the `setup.sh` script in Sandbox it creates a set of
directories and it download OSIE `sandbox/deploy/state/webroot/misc/osie/current`.

    sandbox/deploy/state/webroot/misc/osie/current $ tree -L 1 | clip
    .
    ├── discover-metal-x86_64.tar.gz
    ├── discover-rc
    ├── discover.sh
    ├── grub
    ├── initramfs-2a2
    ├── initramfs-aarch64
    ├── initramfs-amp
    ├── initramfs-hua
    ├── initramfs-qcom
    ├── initramfs-tx2
    ├── initramfs-x86_64
    ├── modloop-2a2
    ├── modloop-aarch64
    ├── modloop-amp
    ├── modloop-hua
    ├── modloop-qcom
    ├── modloop-tx2
    ├── modloop-x86_64
    ├── osie-aarch64.tar.gz
    ├── osie-installer-rc
    ├── osie-installer.sh
    ├── osie-runner-x86_64.tar.gz
    ├── osie-x86_64.tar.gz
    ├── repo-aarch64 -> ../../../alpine/edge
    ├── repo-x86_64 -> ../../../alpine/v3.12
    ├── rescue-helper-rc
    ├── rescue-helper.sh
    ├── runner-rc
    ├── runner.sh
    ├── vmlinuz-2a2
    ├── vmlinuz-aarch64
    ├── vmlinuz-amp
    ├── vmlinuz-hua
    ├── vmlinuz-qcom
    ├── vmlinuz-tx2
    └── vmlinuz-x86_64

    1 directory, 35 files

This is the layout by default. If you want to replace OSIE with noname
you can move `current` to `nop` for example:

    mv ./deploy/state/webroot/misc/osie/current ./deploy/state/webroot/misc/osie/nop

You can download an unpack the new tar

    mkdir current
    wget http://s.gianarb.it/noname/noname-master.tar.gz
    tar xzvf -O ./current ./noname-master.tar.gz

And you are ready to go, now iPXE will boot noname instead of OSIE.

Sandbox has a file called
[current_versions.sh](https://github.com/tinkerbell/sandbox/blob/master/current_versions.sh).
If you change `OSIE_DOWNLOAD_LINK` with the noname link the setup.sh
script will download the OS again and it will uncompress it in
the right location (only if ./deploy/state/webroot/misc/osie/current does not exist)