Don't try to login to quay if a push is made by dependabot
Dependabot PRs are different than other non-contributor PRs in that the branch
comes from this repo, which normally allows the secret to be available for CI
yet the builds still fail (see [1], [2]). It seems that GitHub does not make
secrets available for these PRs as yet another special case [3]. If we skip
logging in to quay for dependabot then we side step this all together.

Dependabot pushes/PRs can be recognized by the github.actor value according to
the GitHub docs [3].

[1]: https://github.com/tinkerbell/hook/pull/89
[2]: https://github.com/tinkerbell/hook/pull/95
[3]: https://docs.github.com/en/code-security/supply-chain-security/keeping-your-dependencies-updated-automatically/automating-dependabot-with-github-actions#handling-pull_request-events

Fixes #92

Signed-off-by: Manuel Mendez <mmendez@equinix.com>