ci: Only push to quay if github.actor is not dependabot
We skipped logging in to quay.io when github.actor is dependabot in [#96] but
did not check if any pushes were attempted. This change will avoid that
failure.

This was missed in the subsequent dependabot PRs [#89] and [#95] because there
was no branch protection setup and I added the ready-to-merge label so mergify
merged it :(. I've added branch protection to require the actions to pass.

[#89]: https://github.com/tinkerbell/hook/pull/89
[#95]: https://github.com/tinkerbell/hook/pull/95
[#96]: https://github.com/tinkerbell/hook/pull/96

Signed-off-by: Manuel Mendez <mmendez@equinix.com>