Overhaul the Makefile
This new setup actually makes use of make's dependency tracking so that
builds aren't randomly broken. It also correctly avoids rebuilds of things
that haven't changed and thus don't need to be rebuilt.

The previous Makefile setup was very buggy/racy if make was run with more
than one job in parallel. It would also have issues if make was interrupted
in the middle of a build. This one does not suffer from any of those issues.

This new setup takes into account both debug vs release modes for the hook.yaml
along with all the arches (new arches should be pretty easy to add in too).

One of the bigger features of this new Makefile setup is that everything
needed to build hook (except for the kernel) is built locally from source. This
means we don't need to push images only to have linuxkit fetch them. This was
only possible previously in dev-mode when we'd load the images from buildx
into the local docker. This was limited to dev mode because docker doesn't
support multi-arch manifests in the local cache. This makes it possible to
tinker locally without the risk of pushing to quay.io unintentionally.

This is also fully parameterized by both TAG and ORG, though ORG isn't
accounted for very well. If ORG is modified in the Makefile please follow
the directions and clean out the build dir `rm -rf out`. The TAG build
parameter is fully supported and can be different between multiple `make`
calls. We make use of this in CI now.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>