ci: Make ci.yaml valid for both pushes and prs
I didn't really see the reason to keep the files split, when ci.yaml is pretty
easily made to work for both scenarios. We end up losing pushes pushing new
container images for non-main branch builds if not created by dependabot. That
was a lot to keep in context and also the correct thing to do. A push to a
branch by someone with write permissions could end up pushing to quay and
confusing things. So all in all we get better determenisitc behavior with
this setup.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>