Add /proc/cmdline parsing for tinkerbell_tls and tink_worker_image:
Boots now has the ability to not pass kernel
commandline args for a registry. When not passing
registry info we will need to know the location of the
tink worker container image. This is handled with the
`tink_worker_image` kernel commandline arg.

Tink server now has the ability to run without TLS.
The `tinkerbell_tls` kernel commandline arg is used to
determine this. This value is parsed and passed along to
the Tink worker.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>