Add syslog docker logger for DinD container:
This sends all DinD container logs (tink-worker, workflow containers)
to the syslog server that Boots currently runs. The kernel commandline
arg `syslog_host` is parsed and used for the syslog destination.

This will allow failed workflow logs to be available in Boots and not
require users to access the machines console.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>