Never bind daemon.json
Linuxkit seems to mount the file as read-only which causes hook-docker to not
be able to overwrite the file. I messed around with adding :rw and converting
to a full `mount` to try and make it rw, but nothing worked. So far we don't
actually need to bind-mount the file so we can just punt on this until later.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>