ci: Publish to GitHub Releases
On both push to main (updates the `latest` release) and tags (creates release
for the tag).

Publishing on main to the latest release gives us a url we can use for sandbox
until a new release is cut or if we want to run sandbox with bleeding edge
hook builds.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>