ci: Limit the kernel push only to branch pushes
Which means we exclude all tags. This is necessary because according to
the docs[1]:

> Path filters are not evaluated for pushes of tags.

So we need to avoid building for tags. This is fine to do because the kernel
isn't versioned by tags but by the contents of the kernel/ tree. I went with
`branches: "**"` instead of `tags-ignore: "**"` because inclusion is easier
to understand vs exclusion imo.

1: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#onpushpull_requestpull_request_targetpathspaths-ignore

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>