Remove address from non tagged interface:
Boots send kernel command line parameter "ip=dhcp",
this causes the system to configure the network interface(s) with DHCP.
When an environment's network configuration has this machine connected a
trunked interface with a default/native VLAN, the interface will be configured
on the default/native VLAN because we haven't yet configured the VLAN interface.
Boots will respond to this DHCP request because in this scenario it is not
VLAN aware. Also in this scenario, the machine will end up being configured
with 2 default routes. To resolve this, we remove the default route and IP
that kernel added and let the dhcpcd handle setting the route.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>