Make linuxkit Hook services resilient so that the getty always comes up:
Hook-bootkit requires a network connection to pull down
the tink-worker image. The previous solution was to make
sure networking was up and running via a DHCP client in the
linuxkit "onboot" section. This meant that any issues with
the DHCP onboot container would cause the getty, and all services,
to never start. This made troubleshooting nearly impossible.

This PR makes Hook-bootkit (and hook-docker) retry all failures.
This allows us to remove the DHCP "onboot" container. This should
allow the getty service to always start.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>