Add os-release file:
This allows us to know at runtime the
version of HookOS. Also, modified the
prompt to contain the version so that
when sharing screenshots to debug it
will be clear the Hook version.

Turned down the kernel.printk log level
to warning (4) so that the console getty
doesn't get messy with kernel messages.
Better UX (hopefully). A side affect of this
is that we won't see messages anymore on the
console for when workflow action containers run.
Syslog messages will still be sent to Boots.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>