Push container images properly:
Because of the way CI is built (2 self-hosted runners),
CI was not pushing container images when building on main.
This was because PR's would build the images and the would
be locally in the Docker cache. Then when main ran,
hook-lk-containers.sh would run and see that the image
existed in the local Docker cache already and stop before
pushing.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>