Add a magic string for the ISO kernel cmdline:
This adds a 1024 character string that is used
as a known string to match on when binary patching
the Hook ISO. This means that a service needs to
binary patch the Hook ISO that is built with the
cmdline. In the Tinkerbell stack this is functionality
that is being built into Smee.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>