Add static network script:
For environments where DHCP is not available,
I've added a script that will statically configure
network interfaces. The IPAM info must be passed in
via kernel cmdline parameters and be in the appropriate
format.

ipam=<mac-address>:<vlan-id>:<ip-address>:<netmask>:<gateway>:<hostname>:<dns>:<search-domains>:<ntp>

This is probably only useful for the HookOS ISO.
For the Tinkerbell stack, Smee handles patching the ISO
at runtime to include this `ipam=` parameter.

To facilitate this static ip configuration, scripts were placed
into the host filesystem at /etc/init.d. Files in this location
are run at startup my the init system. This makes it possible
to just add the scripts as files in the linuxkit yaml file.
The vlan.sh script was moved to an init.d script because it
needs to run before the static-network script. The vlan.sh script
was updated to use posix /bin/sh instead of bash because the host
only has /bin/sh.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>