Rename containerd config, disable cri plugin:
Linuxkit expects the containerd config file to be
at /etc/runtime-config.toml. Rename the config
so that the settings get picked up properly.

Disable the containerd cri plugin. This is not
needed for linuxkit. The cri plugin is a Kubernetes
thing. This removes warning and error messages
that show up and have confused users about real
issues.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>